refactor(npc-ai): split npc_ai along behavior-state seams
Split crates/services/src/cell/service/npc_ai.rs (1692 lines, over the
700-line hard cap) into a npc_ai/ directory along natural behavior-state
seams. Pure code movement — no logic or signature changes.

Layout:
- mod.rs            — module doc, record_decision_outcome, re-exports
- dispatch.rs       — npc_ai_tick state machine + npc_ai_retry_sweep
- fight.rs          — npc_ai_fight + npc_ai_idle_auto_aggro + retry delay
- ability_select.rs — choose_npc_ability, ability_ranges, backup waypoint
- patrol.rs / wander.rs / investigate.rs / follow.rs — movement states
- leash.rs          — leash recovery (snap home + heal)
- lifecycle.rs      — despawn / submit / error terminal states

Re-export discipline: mod.rs pub(super) re-exports npc_ai_tick,
npc_ai_retry_sweep, choose_npc_ability (and the cfg(test) backup-waypoint
shim) so every external `use` path is identical — zero public-surface
change. Moved code's former `super::super::X` (= crate::cell::X) refs
were rewritten to absolute crate:: paths to stay correct one level
deeper.

Behavior-preserving · re-export discipline · zero public-surface change

Part of #529
